Create move_msg_sqs.py
Added new script to move messages from one SQS queue to another
This script will take the source and destination SQS names, a credentials profile and region (default is us_east_1). For now both queues should be in the same region. Will update the script to support Queue in different regions as well.